Update etiquette.md
I suggested changing "One Member's Guidelines" to "General Communication Guidelines."  I took the guidelines in this section and softened the language, while trying to keep the spirit of the guidelines in place.  My goal was to make this section feel more warm and welcoming, as well as feel more helpful.