Clarify and split new committer / new PMC member templates (#183)
Since we are now reaching out to projects about PMC vs. PMC member,
I looked at the source of some of the templates that are used and
clarified, unified them and fixed a few errors in those templates,
also replaced the link to anchor in www-site about new PMC member
with #newpmcmember instead of #newpmc, which could be confusing
and one of the reasons why people use New PMC rather than New PMC
member.

The problem with those templates were that they did not provide
correct templates for inviting (and following up) in case the
PMC member was invited separately from committer - the templates
required manual modifications of Commiter to "PMC member" and
manually modifying text of the template, which could be onoe of
the reasons why people replaced "Committer" with "PMC member" in
their invitations.

The change splits the process out of inviting PMC member separately
(while leaving the optional process where committer is immediately
a PMC member in the original place).

This PR also unifies the way how links are added to the page, there
was a mixture of `[n]` references and `md` way of inlining links.

Another part of this change is introducing "PMC member" everywhere
where "PMC" could be confused with person. For example the
"PMC candidate" - while technically correct ("candidate to the PMC")
is another reason why people could confuse "PMC" with person.
Replacing it with "PMC member candidate" removes the confusion while
being also techincally correct.

The last part of it that in a few places the documentation was
wrongly referring to "asking the board" when inviting PMC member,
where this process have been changed last year to "notifying the board".
All those places have been corrected.